+++ packages/webview_flutter/webview_flutter_android/android/src/main/java/io/flutter/plugins/webviewflutter/GeneratedAndroidWebView.java
@@ -1607,6 +1607,9 @@ public class GeneratedAndroidWebView {
 
     void setTextZoom(@NonNull Long instanceId, @NonNull Long textZoom);
 
+    @Nullable
+    String getUserAgentString(@NonNull Long instanceId);
+
     /** The codec used by WebSettingsHostApi. */
     static @NonNull MessageCodec<Object> getCodec() {
       return new StandardMessageCodec();
@@ -1982,6 +1985,33 @@ public class GeneratedAndroidWebView {
                       (instanceIdArg == null) ? null : instanceIdArg.longValue(),
                       (textZoomArg == null) ? null : textZoomArg.longValue());
                   wrapped.add(0, null);
+                } catch (Throwable exception) {
+                  ArrayList<Object> wrappedError = wrapError(exception);
+                  wrapped = wrappedError;
+                }
+                reply.reply(wrapped);
+              });
+        } else {
+          channel.setMessageHandler(null);
+        }
+      }
+      {
+        BasicMessageChannel<Object> channel =
+            new BasicMessageChannel<>(
+                binaryMessenger,
+                "dev.flutter.pigeon.WebSettingsHostApi.getUserAgentString",
+                getCodec());
+        if (api != null) {
+          channel.setMessageHandler(
+              (message, reply) -> {
+                ArrayList<Object> wrapped = new ArrayList<Object>();
+                ArrayList<Object> args = (ArrayList<Object>) message;
+                Number instanceIdArg = (Number) args.get(0);
+                try {
+                  String output =
+                      api.getUserAgentString(
+                          (instanceIdArg == null) ? null : instanceIdArg.longValue());
+                  wrapped.add(0, output);
                 } catch (Throwable exception) {
                   ArrayList<Object> wrappedError = wrapError(exception);
                   wrapped = wrappedError;