+++ packages/webview_flutter/webview_flutter_android/android/src/test/java/io/flutter/plugins/webviewflutter/WebSettingsTest.java
@@ -4,6 +4,7 @@
 
 package io.flutter.plugins.webviewflutter;
 
+import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
@@ -116,5 +117,11 @@ public class WebSettingsTest {
   public void setTextZoom() {
     testHostApiImpl.setTextZoom(0L, 100L);
     verify(mockWebSettings).setTextZoom(100);
+  }
+
+  @Test
+  public void getWebSettings() {
+    when(mockWebSettings.getUserAgentString()).thenReturn("str");
+    assertEquals(testHostApiImpl.getUserAgentString(0L), "str");
   }
 }