+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.9.0
+
+* Adds support for `PlatformWebViewController.getUserAgent`.
+
 ## 3.8.2
 
 * Fixes unawaited_futures violations.